UPD: HomePageStepper cleanup
